requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch
torchvision
sentence-transformers
requests
pydantic
geopy
pillow
llama-cpp-python